Fixed size_t conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@64199 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seqset/Bioseq_set.cpp
@@ -193,7 +193,7 @@ void CBioseq_set::GetLabel(string* label, ELabelType type) const
             *label += CNcbiOstrstreamToString(os);
             if (this->IsSetSeq_set()) {
                 const TSeq_set& sset = this->GetSeq_set();
-                int len = sset.size();
+                size_t len = sset.size();
                 if (len > 1) {
                     *label += " (" + NStr::SizetToString(sset.size()) + " components)";
                 } else if (len == 1) {
